+++ frontend/zigbee2mqtt.js
@@ -159,7 +159,9 @@ function(app, Viz, vizRenderer, leaflet) {
                         model: null,
                     }, device, {
                         lastSeen: device.lastSeen || 'N/A',
-                        description: device.description || ''
+                        description: device.description || '',
+                        model: device.model || 'N/A',
+                        type: device.type || 'N/A'
                     });
                 }).sort(function(a, b) {
                     return a.friendly_name < b.friendly_name ? -1 : 1